pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hr-screening"
version = "0.1.0"
description = "Automated candidate resume screening core and adapters."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Lazuli Hiring Automation" }]
dependencies = [
    "pydantic>=2.6,<3.0",
    "pendulum>=3.0,<4.0",
    "python-dateutil>=2.9",
    "structlog>=24.1",
    "dependency-injector>=4.41",
    "pluggy>=1.5",
    "orjson>=3.10",
    "pandas>=2.2",
    "rapidfuzz>=3.9",
    "regex>=2024.5.15",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "hypothesis>=6.87",
    "ruff>=0.5.0",
    "black>=24.4",
    "mypy>=1.10",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7",
    "types-python-dateutil",
    "types-requests",
]
pdf = [
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "pypdf>=4.2",
]
nlp = [
    "sudachipy>=0.6.8",
    "sudachidict-core>=20240409",
]
ml = [
    "scikit-learn>=1.4",
    "sentence-transformers>=3.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q --cov=hrscreening --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["hrscreening"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict = true
